refactor: make ParamAwareCommandTrait into AbstractParamAwareCommand
This patch converts `ParamAwareCommandTrait` into the abstract class
`AbstractParamAwareCommand`, which makes it possible for end-users to
override the `run()` method in their own implementations if needed.

To make this possible, the patch also introduces the `final` method
`decorateInputToBeParamAware()`, which accepts the input and output, and
uses them to return a `ParamAwareInputInterface` implementation. This
allows users overriding `run()` to do either of:

```php
parent::run($input, $output)
```

or

```php
$input = $this->decorateInputToBeParamAware($input, $output);
```

If you want a command to be parameter aware, it now MUST extend
`AbstractParamAwareCommand`.
