+++ test/TestAsset/ParamAwareCommandStub.php
@@ -10,10 +10,12 @@ declare(strict_types=1);
 
 namespace LaminasTest\Cli\TestAsset;
 
+use Laminas\Cli\Command\AbstractParamAwareCommand;
+use Symfony\Component\Console\Helper\HelperSet;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class ParamAwareCommandStub
+class ParamAwareCommandStub extends AbstractParamAwareCommand
 {
     /** @var InputInterface */
     public $input;
@@ -21,7 +23,52 @@ class ParamAwareCommandStub
     /** @var OutputInterface */
     public $output;
 
-    public function run(InputInterface $input, OutputInterface $output): int
+    /** @var array */
+    public $options = [];
+
+    /** HelperSet */
+    private $helperSet;
+
+    public function __construct(HelperSet $helperSet)
+    {
+        $this->helperSet = $helperSet;
+        parent::__construct('test');
+    }
+
+    /**
+     * @param string|array|null $shortcut
+     * @param null|mixed        $default Defaults to null.
+     * @return $this
+     */
+    public function addOption(
+        string $name,
+        $shortcut = null,
+        ?int $mode = null,
+        string $description = '',
+        $default = null
+    ) {
+        $this->options[$name] = [
+            'shortcut'    => $shortcut,
+            'mode'        => $mode,
+            'description' => $description,
+            'default'     => $default,
+        ];
+        return $this;
+    }
+
+    /**
+     * @return HelperSet
+     */
+    public function getHelperSet()
+    {
+        return $this->helperSet;
+    }
+
+    protected function configure(): void
+    {
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $this->input  = $input;
         $this->output = $output;
